+++ hug/test.py
@@ -60,7 +60,7 @@ def call(method, api_or_module, url, body='', headers=None, params=None, query_s
             data = data.getvalue()
             try:
                 response.data = data.decode('utf8')
-            except UnicodeDecodeError:
+            except UnicodeDecodeError:   # pragma: no cover
                 response.data = data
         except (UnicodeDecodeError, AttributeError):
             response.data = result[0]