+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog
 - Added built in handlers for CORS support:
     - directive `hug.directives.cors`
     - Improved routing support
+    - Added allow origins middleware
 
 ### 2.3.0 - May 4, 2017
 - Falcon requirement upgraded to 1.2.0